Run Luminescence shiny apps

A wrapper for shiny::runApp to start interactive shiny apps for the R package Luminescence.

The RLumShiny package provides a single function from which all shiny apps can be started: app_RLum(). It essentially only takes one argument, which is a unique keyword specifying which application to start. See the table below for a list of available shiny apps and which keywords to use. If no keyword is used a dashboard will be started instead, from which an application can be started.

Application name:	Keyword:	Function:

Abanico Plot	abanico	Luminescence::plot_AbanicoPlot

Histogram	histogram	Luminescence::plot_Histogram

Kernel Density Estimate Plot	KDE	Luminescence::plot_KDE

Radial Plot	radialplot	Luminescence::plot_RadialPlot

Aliquot Size	aliquotsize	Luminescence::calc_AliquotSize

Dose Recovery Test	doserecovery	Luminescence::plot_DRTResults

Cosmic Dose Rate	cosmicdose	Luminescence::calc_CosmicDoseRate

CW Curve Transformation	transformCW	Luminescence::convert_CW2pHMi, Luminescence::convert_CW2pLM, Luminescence::convert_CW2pLMi, Luminescence::convert_CW2pPMi

Filter Combinations	filter	Luminescence::plot_FilterCombinations

Fast Ratio	fastratio	Luminescence::calc_FastRatio

Fading Correction	fading	Luminescence::analyse_FadingMeasurement, Luminescence::calc_FadingCorr

Finite Mixture	finitemixture	Luminescence::calc_FiniteMixture

Huntley (2006)	huntley2006	Luminescence::calc_Huntley2006

IRSAR RF	irsarRF	Luminescence::analyse_IRSAR.RF

LM Curve	lmcurve	Luminescence::fit_LMCurve

Test Stimulation Power	teststimulationpower	Luminescence::plot_RLum

Scale Gamma Dose Rate	scalegamma	Luminescence::scale_GammaDose

RCarb app	RCarb	RCarb::model_DoseRate

The app_RLum() function is just a wrapper for shiny::runApp. Via the ... argument further arguments can be directly passed to shiny::runApp. See ?shiny::runApp for further details on valid arguments.

Usage

app_RLum(app = NULL, ...)

Arguments

app

character (required): name of the application to start. See details for a list of available apps.

...

further arguments to pass to shiny::runApp

See also

Author

Christoph Burow, University of Cologne (Germany)

Examples

if (FALSE) { # \dontrun{
# Dashboard
app_RLum()

# Plotting apps
app_RLum("abanico")
app_RLum("histogram")
app_RLum("KDE")
app_RLum("radialplot")
app_RLum("doserecovery")

# Further apps
app_RLum("aliquotsize")
app_RLum("cosmicdose")
app_RLum("transformCW")
app_RLum("filter")
app_RLum("fastratio")
app_RLum("fading")
app_RLum("finitemixture")
app_RLum("huntley2006")
app_RLum("irsarRF")
app_RLum("lmcurve")
app_RLum("surfaceexposure")
app_RLum("teststimulationpower")
app_RLum("scalegamma")
app_RLum("RCarb")
} # }