app_RLum {RLumShiny}	R Documentation

Run Luminescence shiny apps

Description

A wrapper for runApp to start interactive shiny apps for the R package Luminescence.

Usage


app_RLum(app = NULL, ...)

Arguments

app
character (required): name of the application to start. See details for a list of available apps.

...
further arguments to pass to runApp

Details

The RLumShiny package provides a single function from which all shiny apps can be started: app_RLum(). It essentially only takes one argument, which is a unique keyword specifying which application to start. See the table below for a list of available shiny apps and which keywords to use. If no keyword is used a dashboard will be started instead, from which an application can be started.

Application name: 	 Keyword: 	 Function:

Abanico Plot 	 abanico 	 plot_AbanicoPlot

Histogram 	 histogram 	 plot_Histogram

Kernel Density Estimate Plot 	 KDE 	 plot_KDE

Radial Plot 	 radialplot 	 plot_RadialPlot

Dose Recovery Test 	 doserecovery 	 plot_DRTResults

Cosmic Dose Rate 	 cosmicdose 	 calc_CosmicDoseRate

CW Curve Transformation 	 transformCW 	 CW2pHMi, CW2pLM, CW2pLMi, CW2pPMi

Filter Combinations 	 filter 	 plot_FilterCombinations

Fast Ratio 	 fastratio 	 calc_FastRatio

Fading Correction 	 fading 	 analyse_FadingMeasurement, calc_FadingCorr

Test Stimulation Power 	 teststimulationpower 	 plot_RLum

Scale Gamma Dose Rate 	 scalegamma 	 scale_GammaDose

RCarb app 	 RCarb 	 RCarb::model_DoseRate

The app_RLum() function is just a wrapper for runApp. Via the ... argument further arguments can be directly passed to runApp. See ?shiny::runApp for further details on valid arguments.

Author(s)

Christoph Burow, University of Cologne (Germany)

See Also

runApp

Examples



## Not run: 
# Dashboard
app_RLum()

# Plotting apps
app_RLum("abanico")
app_RLum("histogram")
app_RLum("KDE")
app_RLum("radialplot")
app_RLum("doserecovery")

# Further apps
app_RLum("cosmicdose")
app_RLum("transformCW")
app_RLum("filter")
app_RLum("fastratio")
app_RLum("fading")
app_RLum("surfaceexposure")
app_RLum("teststimulationpower")
app_RLum("scalegamma")
app_RLum("RCarb")

## End(Not run)


[Package RLumShiny version 0.2.2 Index]