Create a JSColor picker input widget

Creates a JSColor (Javascript/HTML Color Picker) widget to be used in shiny applications.

jscolorInput(inputId, label, value, position = "bottom",
  color = "transparent", mode = "HSV", slider = TRUE,
  close = FALSE)

Arguments

inputId
character (required): Specifies the input slot that will be used to access the value.

label
character (optional): Display label for the control, or NULL for no label.

value
character (optional): Initial RGB value of the color picker. Default is black ('#000000').

position
character (with default): Position of the picker relative to the text input ('bottom', 'left', 'top', 'right').

color
character (with default): Picker color scheme ('transparent' by default). Use RGB color coding ('000000').

mode
character (with default): Mode of hue, saturation and value. Can either be 'HSV' or 'HVS'.

slider
logical (with default): Show or hide the slider.

close
logical (with default): Show or hide a close button.

See also

Examples

# html code
jscolorInput("col", "Color", "21BF6B", slider = FALSE)

#> <p>Color</p>
#> <input id="col" value="21BF6B" class="color {hash:true, pickerPosition:&#39;bottom&#39;, pickerBorderColor:&#39;transparent&#39;, pickerFaceColor:&#39;transparent&#39;, pickerMode:&#39;HSV&#39;, slider:false, pickerClosable:false} shiny-bound-input" onchange="$(&#39;#col&#39;).trigger(&#39;afterChange&#39;)"/>
#> <script>$('#col').trigger('afterChange')</script>

# example app

# NOT RUN {
shinyApp(
ui = fluidPage(
  jscolorInput(inputId = "col", label = "JSColor Picker",
               value = "21BF6B", position = "right",
               mode = "HVS", close = TRUE),
  plotOutput("plot")
),
server = function(input, output) {
  output$plot <- renderPlot({
    plot(cars, col = input$col, cex = 2, pch = 16)
 })
})
# }